Added subscriptions recurring functions

+++ needly_utils.py
@@ -1,6 +1,6 @@
 from datetime import datetime, time, timedelta
 from peewee import fn
-from schema import Chat, User, Needly
+from schema import Chat, Job, User, Needly
 from secret import TM_TOKEN
 from binance.spot import Spot as Client
 import tradermade as tm
@@ -8,6 +8,10 @@ import random
 import json
 import calendar
 import requests
+import pytz
+
+# Define Singapore time zone
+SGT = pytz.timezone('Asia/Singapore')
 
 
 def register_user(chat_id: int, user_id: int) -> bool:
@@ -18,44 +22,50 @@ def register_user(chat_id: int, user_id: int) -> bool:
 
 
 def reset_needly_display_order():
-    chats = Chat.select()
-    for chat in chats:
-        # Fetch all Needly records for a chat, ordered by the current ID or another relevant field
-        needly_records = Needly.select().where(Needly.chat == chat).order_by(Needly.date.asc())
-
-        # Update each record with a new display_id in sequential order
-        for index, record in enumerate(needly_records, start=1):
-            record.display_id = index
-            record.save()  # Save the update to the database
+    # Fetch all entries ordered by created_at
+    all_entries = Needly.select().order_by(Needly.created_at)
+    # Assign display_id sequentially
+    for idx, record in enumerate(all_entries, start=1):
+        record.display_id = idx
+        record.save()
 
 
-def log_entry(chat_id: int, user_id: int, log_cat: str, log_amt: float, log_desc: str, date: datetime) -> Needly:
-    chat, _ = Chat.get_or_create(chat_id=chat_id)
-    user, _ = User.get_or_create(user_id=user_id)
-
-    log = Needly.create(chat=chat, user=user, log_cat=log_cat.upper(), log_amt=log_amt,
-                        log_desc=log_desc, date=date)
-    
-    reset_needly_display_order() # Reset the display order after deleting a record
-
-    return log
+def log_entry(chat_id, user_id, log_cat, log_amt, log_desc, date):
+    # Create a new Needly entry
+    needly_entry = Needly.create(
+        chat=Chat.get(Chat.chat_id == chat_id),
+        user=User.get(User.user_id == user_id),
+        log_cat=log_cat,
+        log_amt=log_amt,
+        log_desc=log_desc,
+        date=date
+    )
+    reset_needly_display_order()  # Reset the display order after logging an entry
+    return needly_entry
 
 
 def delete_entry_db(del_id: int) -> tuple:
-    delete = Needly.delete().where(Needly.display_id==del_id).execute()
-    # delete = Needly.delete().where(del_id).execute()
+    needly_entry = Needly.get_or_none(Needly.display_id == del_id)
 
-    reset_needly_display_order() # Reset the display order after deleting a record
+    if needly_entry:
+        delete = Needly.delete().where(Needly.display_id == del_id).execute()
+        reset_needly_display_order()  # Reset the display order after deleting a record
+        return delete
+    else:
+        return 0
 
-    return delete
+
+def get_all_jobs():
+    return Job.select()
 
 
 def db_get_by_id(display_id: int):
-    requested_entry = Needly.select(Needly.log_cat, Needly.log_amt, Needly.log_desc).where(Needly.display_id == display_id)
+    requested_entry = Needly.select(Needly.log_cat, Needly.log_amt, Needly.log_desc).where(
+        Needly.display_id == display_id)
 
     result = []
 
-    #0 = Description, #1 = Amount, #2 = Category
+    # 0 = Description, #1 = Amount, #2 = Category
     for entry in requested_entry:
         result.append(entry.log_desc)
         result.append(entry.log_amt)
@@ -69,7 +79,7 @@ def dict_months(chat_id: int, user_id: int) -> dict:
     user, _ = User.get_or_create(user_id=user_id)
 
     if chat_created:
-        return []
+        return {}
 
     chat_month = Needly.select(
         Needly.display_id,
@@ -82,11 +92,23 @@ def dict_months(chat_id: int, user_id: int) -> dict:
     month_dict = {}
 
     for entry in chat_month:
+        # Ensure entry.date is a datetime object
+        if isinstance(entry.date, str):
+            try:
+                entry.date = datetime.strptime(
+                    entry.date, '%Y-%m-%d %H:%M:%S.%f%z')
+            except ValueError:
+                entry.date = datetime.strptime(
+                    entry.date, '%Y-%m-%d %H:%M:%S%z')
+
+        entry.date = entry.date.astimezone(pytz.timezone('Asia/Singapore'))
+
         date_string = entry.date.strftime('%d-%m-%Y')
         date = date_string.split('-')
         get_date, get_month, get_year = date[0], date[1], date[2]
 
-        get_date, get_month, get_year = int(get_date), int(get_month), int(get_year)
+        get_date, get_month, get_year = int(
+            get_date), int(get_month), int(get_year)
 
         if get_year not in month_dict:
             month_dict[get_year] = {}
@@ -156,36 +178,36 @@ def get_month_dates(month_name, year):
 def get_weeks_in_month(year, month):
     # Get the first day of the month
     first_day = datetime(year, month, 1)
-    
+
     # Get the last day of the month
     if month == 12:
         last_day = datetime(year, month, 31)
     else:
         last_day = datetime(year, month + 1, 1) - timedelta(days=1)
-    
+
     # Initialize a list to store the weeks
     weeks = []
 
     # Loop through the weeks of the month
     current_week = []
     current_day = first_day
-    
+
     while current_day <= last_day:
         # Add the day and date to the current week
         current_week.append((current_day.strftime('%A'), current_day.day))
-        
+
         # If it's the last day of the week, start a new week
         if current_day.weekday() == 6:
             weeks.append(current_week)
             current_week = []
-        
+
         # Move to the next day
         current_day += timedelta(days=1)
-    
+
     # Add the last week if it's not empty
     if current_week:
         weeks.append(current_week)
-    
+
     return weeks
 
 
@@ -207,10 +229,10 @@ def fetch_data(year, month, chat_id, user_id) -> dict:
     calendar_emoji = u"\U0001F4C5"
     stock_rising = u'\U0001F4C8'
 
-    fun_replies = ["No expenses logged on this day!", 
-                   "Wow that's surprising, you didn't spend any money", 
-                   "No way you didn't spend any money...", 
-                   "That's cap", 
+    fun_replies = ["No expenses logged on this day!",
+                   "Wow that's surprising, you didn't spend any money",
+                   "No way you didn't spend any money...",
+                   "That's cap",
                    "No money flowed out of your account?!",
                    "Wow that's a first, not a single cent spent"]
 
@@ -220,7 +242,7 @@ def fetch_data(year, month, chat_id, user_id) -> dict:
         week_total = 0.0
         if i not in res_dict:
             res_dict[i] = f"*Week {i}/{len(weeks)} of {month_name}{calendar_emoji}*\n\n"
-        
+
         for day, date in week:
             day_total = 0.0
             res_dict[i] += f"*{date} {month_name} ({day})*\n"
@@ -230,7 +252,7 @@ def fetch_data(year, month, chat_id, user_id) -> dict:
                 if datetime(year, month, date) > datetime.now():
                     res_dict[i] += f"You have yet to log a future expense! {thinking_face}\n\n"
 
-                else: 
+                else:
                     rng = random.randint(0, len(fun_replies) - 1)
                     res_dict[i] += f"{fun_replies[rng]} {thinking_face}\n\n"
 
@@ -246,9 +268,9 @@ def fetch_data(year, month, chat_id, user_id) -> dict:
                     day_total += amount
 
                     res_dict[i] += f"*[ID. {id}]* {description}: {amount:.2f} *[{cat}]*\n"
-                
+
                 res_dict[i] += f"{money_fly}*Daily Total = {day_total:.2f}*\n\n"
-            
+
             week_total += day_total
 
         res_dict[i] += f"{money_bag}*Weekly Total = {week_total:.2f}*\n"
@@ -262,15 +284,16 @@ def fetch_data(year, month, chat_id, user_id) -> dict:
 def get_forex(ticker):
     tm.set_rest_api_key(TM_TOKEN)
 
-    request_currency = tm.live(currency=ticker, fields = ["mid"]).at[0, "mid"]
+    request_currency = tm.live(currency=ticker, fields=["mid"]).at[0, "mid"]
     return request_currency
 
 
 def get_btc_price():
     # Send request to CoinDesk API
-    response = requests.get('https://api.coindesk.com/v1/bpi/currentprice.json')
+    response = requests.get(
+        'https://api.coindesk.com/v1/bpi/currentprice.json')
     data = response.json()
-    
+
     # Access the Bitcoin rate in USD
     bitcoin_rate_str = data["bpi"]["USD"]["rate"].replace(',', '')
     bitcoin_rate_float = float(bitcoin_rate_str)
@@ -284,7 +307,7 @@ def get_btc_price():
 def get_crypto_price(tickers):
     try:
         spot_client = Client(base_url="https://testnet.binance.vision")
-        
+
         if len(tickers) == 1:
             response = spot_client.ticker_price(symbol=tickers[0])
         else:
@@ -294,64 +317,3 @@ def get_crypto_price(tickers):
         return None
 
     return response
-
-
-############################################################################################################
-# Deprecated Functions
-""" def filter_week(chat_id: int, user_id: int) -> dict:
-    chat, chat_created = Chat.get_or_create(chat_id=chat_id)
-    user, _ = User.get_or_create(user_id=user_id)
-    start_of_week = datetime.now() - timedelta(days=datetime.now().weekday())
-    start_of_week = start_of_week.strftime('%d-%m-%Y')
-    start_of_week = datetime.strptime(start_of_week, '%d-%m-%Y')
-
-    if chat_created:
-        return []
-
-    chat_month = Needly.select(
-        Needly.display_id,
-        Needly.log_cat,
-        Needly.log_amt,
-        Needly.log_desc,
-        Needly.date).where(
-            (Needly.chat == chat) &
-            (Needly.date.between(start_of_week, datetime.now()))
-        )
-
-    month_dict = {}
-
-    for entry in chat_month:
-        date_string = entry.date.strftime('%d-%m-%Y')
-        date = date_string.split('-')
-        get_date, get_month, get_year = date[0], date[1], date[2]
-
-        get_date, get_month, get_year = int(get_date), int(get_month), int(get_year)
-
-        if get_year not in month_dict:
-            month_dict[get_year] = {}
-
-        if get_month not in month_dict[get_year]:
-            month_dict[get_year][get_month] = {}
-
-        if get_date not in month_dict[get_year][get_month]:
-            month_dict[get_year][get_month][get_date] = []
-
-        month_dict[get_year][get_month][get_date].append(
-            (entry.id, entry.log_desc, entry.log_amt, entry.log_cat))
-
-    return month_dict """
-
-
-""" def within_week(date) -> bool:
-    within_week = False
-    start_of_week = datetime.now() - timedelta(days=datetime.now().weekday())
-    if 0 <= int(date) - start_of_week.day <= 7 or -23 <= int(date) - start_of_week.day <= -26:
-        within_week = True
-        return within_week
-
-    elif -26 <= int(date) - start_of_week.day <= -23:
-        within_week = True
-        return within_week
-
-    return within_week """
-############################################################################################################
